Adjust hostnames to match the container name
As this is running in the docker network, the hostnames used to connect here should the container name. This alleviates the need for any `/etc/hosts` modification. 

curl already has a `-k` to handle the untrusted CA, the cert name mismatch won't really matter. 

Seems to work on my laptop